Removed unwanted dependent packages and created a query service layer for query service functionalities
